refactor winstreaks
