+++ src/renderer/i18n/es-ES.json
@@ -9,6 +9,7 @@
         "character": "Character | Characters",
         "boss": "Boss | Bosses",
         "remove": "Remove",
+        "edit": "Edit",
         "cancel": "Cancel",
         "close": "Close",
         "add": "Add",
