+++ src/renderer/i18n/fr-FR.json
@@ -9,6 +9,7 @@
         "character": "Personnage | Personnages",
         "boss": "Boss | Bosses",
         "remove": "Effacer",
+        "edit": "Editer",
         "cancel": "Annuler",
         "close": "Fermer",
         "add": "Ajouter",
