+++ src/renderer/i18n/ja-JP.json
@@ -9,6 +9,7 @@
         "character": "Character | Characters",
         "boss": "Boss | Bosses",
         "remove": "Remove",
+        "edit": "Editer",
         "cancel": "Cancel",
         "close": "Close",
         "add": "Add",
